career-assistant-flask/requirements.txt
Flask
python-dotenv
requests
PyPDF2
langchain
langchain-google-genai
langchain-community
langchain-chroma
google-generativeai
